New implementation for working with retrofit, Room and Networking.
Uses dependency injection to inject network and retrofit functions into the screen composable.

Removed primary key id and replaced it with latitude and longitude.

Implemented a network manager for dealing with network related problems.

Implemented a new Retrofit class.

Added a new class CustomStringResourcesClass.kt for working with string parsing with date and for working with latitude and longitude strings.