Fixed screenupdate, added exceptions when inputs are not float.
Added a Snackbar host for presenting a message when the latitude and longitude values are not float.

Changed the screen to present the equivalent to what the instructions for lab 1.2 is showing.

Added a CustomComposable for displaying Icons.

Still trying to figure out how to extract error message from retrofit if it fails. Currently, the app does not crash when out of bounds coordinates are used.