Removed TODO notations.
Fixed bug where the ui would not update unless the scrollable list was scrolled for a few seconds.
The downside of this approach is that the contents of the scrollable list will first be empty (the same way as when the app is started for the first time), and then the scrollable list is filled with the current weather reports.

Fixed bug where the first press of the set button would not yield weather report.